Added search page, fixed the redirect after login and register, added validation
Things to do:
- fix the redirect after add or update
- using cookies

+++ routes/web.php
@@ -21,7 +21,8 @@ use Illuminate\Support\Facades\Route;
 */
 
 // INDEX PAGE
-Route::get('/', [ProductController::class, 'index']);
+// Route::get('/', [ProductController::class, 'index']);
+Route::get('/', [HomeController::class, 'index']);
 // Route::get('/member-index', [HomeController::class, 'memberIndex'])->middleware('role:member');
 // Route::get('/admin-index', [HomeController::class, 'adminIndex'])->middleware('role:admin');
 
@@ -35,36 +36,34 @@ Route::get('/login', [PagesController::class, 'login']);
 Route::get('/register', [PagesController::class, 'register']);
 
 // VIEW, ADD, UPDATE, DELETE PRODUCT
-Route::get('/view-product', [ProductController::class, 'displayAll']);
-Route::get('/add-product', [ProductController::class, 'getAddProductPage']);
-Route::post('/add-product', [ProductController::class, 'addProduct']);
-Route::post('/delete-product/{id}', [ProductController::class, 'deleteProduct']);
-Route::get('/edit-product/{id}', [ProductController::class, 'getEditProductPage']);
-Route::post('/edit-product/{id}', [ProductController::class, 'editProduct']);
+Route::get('/view-product', [ProductController::class, 'displayAll'])->middleware('role:admin');
+Route::get('/add-product', [ProductController::class, 'getAddProductPage'])->middleware('role:admin');
+Route::post('/add-product', [ProductController::class, 'addProduct'])->middleware('role:admin');
+Route::post('/delete-product/{id}', [ProductController::class, 'deleteProduct'])->middleware('role:admin');
+Route::get('/edit-product/{id}', [ProductController::class, 'getEditProductPage'])->middleware('role:admin');
+Route::post('/edit-product/{id}', [ProductController::class, 'editProduct'])->middleware('role:admin');
 
 // VIEW, ADD, UPDATE, DELETE CATEGORY
-Route::get('/view-category', [CategoryController::class, 'displayAll']);
-Route::get('/add-category', [CategoryController::class, 'getAddCategoryPage']);
-Route::post('/add-category', [CategoryController::class, 'addCategory']);
-Route::post('/delete-category/{id}', [CategoryController::class, 'deleteCategory']);
-Route::get('/edit-category/{id}', [CategoryController::class, 'getEditCategoryPage']);
-Route::post('/edit-category/{id}', [CategoryController::class, 'editCategory']);
+Route::get('/view-category', [CategoryController::class, 'displayAll'])->middleware('role:admin');
+Route::get('/add-category', [CategoryController::class, 'getAddCategoryPage'])->middleware('role:admin');
+Route::post('/add-category', [CategoryController::class, 'addCategory'])->middleware('role:admin');
+Route::post('/delete-category/{id}', [CategoryController::class, 'deleteCategory'])->middleware('role:admin');
+Route::get('/edit-category/{id}', [CategoryController::class, 'getEditCategoryPage'])->middleware('role:admin');
+Route::post('/edit-category/{id}', [CategoryController::class, 'editCategory'])->middleware('role:admin');
 
 // CART
-Route::get('/view-cart/{id}', [CartController::class, 'displayAll']);
-Route::post('/add-cart-item', [CartDetailController::class, 'addItem']);
-Route::post('/delete-cart-item/{cart_id}/{product_id}', [CartDetailController::class, 'deleteItem']);
-Route::get('/update-cart-item/{cart_id}/{product_id}', [CartDetailController::class, 'getUpdateItemPage']);
-Route::post('/update-cart-item/{cart_id}/{product_id}', [CartDetailController::class, 'updateItem']);
+Route::get('/view-cart/{id}', [CartController::class, 'displayAll'])->middleware('role:member');
+Route::post('/add-cart-item', [CartDetailController::class, 'addItem'])->middleware('role:member');
+Route::post('/delete-cart-item/{cart_id}/{product_id}', [CartDetailController::class, 'deleteItem'])->middleware('role:member');
+Route::get('/update-cart-item/{cart_id}/{product_id}', [CartDetailController::class, 'getUpdateItemPage'])->middleware('role:member');
+Route::post('/update-cart-item/{cart_id}/{product_id}', [CartDetailController::class, 'updateItem'])->middleware('role:member');
 
 // TRANSACTION
-Route::post('/checkout/{cart_id}', [TransactionController::class, 'checkout']);
+Route::post('/checkout/{cart_id}', [TransactionController::class, 'checkout'])->middleware('role:member');
+Route::get('/history-transaction', [TransactionController::class, 'displayAll'])->middleware('role:member');
 
-// Route::get('/history-transaction', function () {
-//     return view('history-transaction');
-// });
-
-Route::get('/history-transaction', [TransactionController::class, 'displayAll']);
+// SEARCH
+Route::get('/search/{page?}', [HomeController::class, 'searchProducts']);
 
 // AUTH
 Auth::routes();
